добавил 3 экран ui

+++ app/src/main/java/com/example/findequalphoto/ui/MainVM.kt
@@ -38,9 +38,13 @@ class MainVM(private val photoRepo: PhotoRepo) : ViewModel() {
         }
     }
 
-    fun deletePhoto(){
+    fun deletePhoto(isFinish:Boolean=false){
         viewModelScope.launch {
-            photoRepo.deleteAllPhoto()
+            if(!isFinish) {
+                photoRepo.deleteAllPhoto()
+            }else{
+                photoRepo.finishProgress()
+            }
         }
     }
